isolate inline_asm. Once compiler ready, it should be removed.
